Add not_if and only_if to Script pattern.
This allows you to more easily skip actions once they have already run.
I've found that to be a valuable optimization and avoids somewhat
problematic cases where installing a package after it has already been
installed to cause an upgrade.

I'm not sure if this is the best way to implement it but it seemed like
it should work. If there is a better way I'm open to suggestions.